requirements.txt
numpy
pandas
scikit-learn
sentence-transformers
torch
faiss-cpu
fastapi
uvicorn
pyyaml
tqdm
pytest
